requirements.txt
langchain>=0.1.20
langgraph>=0.0.40
openai>=1.3.7
faiss-cpu
python-dotenv
langchain-openai
